refactor: split shared/types.ts into domain-specific modules
Break the 976-line monolithic types file into 17 focused modules under
shared/types/ (projects, agents, sessions, algochat, councils, memories,
work-tasks, credits, schedules, webhooks, workflows, voice, personas,
skill-bundles, a2a, reputation, marketplace). Original file re-exports
from the barrel index for full backward compatibility — all 56 importing
files continue to work unchanged.

Closes #444
