+++ shared/types/memories.ts
@@ -1,0 +1,12 @@
+export type MemoryStatus = 'pending' | 'confirmed' | 'failed';
+
+export interface AgentMemory {
+    id: string;
+    agentId: string;
+    key: string;
+    content: string;
+    txid: string | null;
+    status: MemoryStatus;
+    createdAt: string;
+    updatedAt: string;
+}
