+++ shared/types/voice.ts
@@ -1,0 +1,3 @@
+export type NotificationChannelType = 'websocket' | 'discord' | 'telegram' | 'github' | 'algochat' | 'whatsapp' | 'signal' | 'slack';
+
+export type VoicePreset = 'alloy' | 'echo' | 'fable' | 'onyx' | 'nova' | 'shimmer';
